refactor(project): reorganize router and query client structure
- move `routeTree.gen.ts` to `src/` root
- move `router` and `query-client` modules from `src/lib/` to `src/app/`
- update all build, lint, and IDE configuration files to reflect path changes
