+++ .autocorrectignore
@@ -3,4 +3,4 @@
 **/yarn.lock
 **/*.min.*
 **/CHANGELOG.md
-src/lib/router/routeTree.gen.ts
+src/routeTree.gen.ts
