+++ docs/conventions.md
@@ -78,7 +78,8 @@ src/
 ├── styles/
 ├── testing/
 ├── utils/
-└── index.tsx
+├── index.tsx
+└── routeTree.gen.ts
 ```
 
 - `@types/`: global type declarations and tool-generated declarations.
@@ -101,6 +102,7 @@ src/
 - `testing/`: test utilities, custom render helpers, and test setup helpers.
 - `utils/`: generic utility functions that can be reused across projects.
 - `index.tsx`: React bootstrap entry point.
+- `routeTree.gen.ts`: TanStack Router-generated application route tree.
 
 ## Naming
