+++ src/app/providers/router.tsx
@@ -1,6 +1,6 @@
 import { RouterProvider } from '@tanstack/react-router';
 
-import { router } from '@/lib/router';
+import { router } from '../router';
 
 export function AppRouter() {
   return <RouterProvider router={router} />;
